Working solution with rBCG and rPPG in two files.

+++ face_recognition/combinedSolution.cpp
@@ -4,10 +4,9 @@ int main(int argc, const char** argv) {
 
 	//introduce the module
 	CommandLineParser parser(argc, argv,
-		"{help h||}"
 		"{face_cascade|data/haarcascades/haarcascade_frontalface_alt.xml|Path to face cascade.}"
 		"{camera|0|Camera device number.}");
-	parser.about("\nPress Esc to quit program\nThis program demonstrates signal decomposition by color in openCV with facial recognition in a video stream");
+	parser.about("\nPress Esc to quit program\nThis program demonstrates signal decomposition by color and motion in openCV with facial recognition in a video stream");
 	parser.printMessage();
 
 	//-- 1. Load the cascades
@@ -16,16 +15,30 @@ int main(int argc, const char** argv) {
 		std::cout << "--(!)Error loading face cascade\n";
 		return -1;
 	};
-	//Get video input/output setup
+
+	std::cout << "Please select the correct configuration for your camera: " << std::endl;
+	std::cout << "Enter '1' for 1920*1080, or '2' for 1280*720" << std::endl;
+	int choice;
+	std::cin >> choice;
+
 	cv::VideoCapture capture(1); // 1 for front facing webcam
 	capture.set(CAP_PROP_FPS, 30);
-	capture.set(CAP_PROP_FRAME_WIDTH, 1920);
-	capture.set(CAP_PROP_FRAME_HEIGHT, 1080);
-	capture.set(CAP_PROP_AUTO_EXPOSURE, 0);
 	capture.set(CAP_PROP_AUTOFOCUS, 1);
+	capture.set(CAP_PROP_AUTO_EXPOSURE, 0);
 	capture.set(CAP_PROP_AUTO_WB, 0);
+	if (choice == 1) {
+		//Get video input/output setup
+		capture.set(CAP_PROP_FRAME_WIDTH, 1920);
+		capture.set(CAP_PROP_FRAME_HEIGHT, 1080);
+	}
+	else {
+		capture.set(CAP_PROP_FRAME_WIDTH, 1280);
+		capture.set(CAP_PROP_FRAME_HEIGHT, 720);
+	}
 
 	std::cout << "CAPTURE FOMRAT IS: " << capture.get(CAP_PROP_FORMAT) << std::endl;
+
+	//checks to see if the camera works for both methods
 	if (!capture.isOpened()) {
 		capture.release();
 		capture.open(0);
@@ -44,7 +57,8 @@ int main(int argc, const char** argv) {
 	bool recording = true;
 	bool processing = false;
 	bool initialising = true;
-
+	bool plotLegends = true;
+	bool firstRecording = true;
 	//Live-plotting (non-blocking operation of plotting)
 	plt::figure_size(1000, 800);
 	plt::show(false);
@@ -69,10 +83,10 @@ int main(int argc, const char** argv) {
 		cv::Mat frame;
 		if (recording && !processing) {
 			if (initialising) {
-				numFrames = 30 * 5;
+				numFrames = 33 * 5;
 			}
 			else {
-				numFrames = 30 * 15;
+				numFrames = 33 * 15;
 			}
 			for (int i = 0; i < numFrames; i++) {
 				Clock::time_point start = Clock::now();
@@ -81,9 +95,16 @@ int main(int argc, const char** argv) {
 				}
 				Clock::time_point end = Clock::now();
 				auto ms = std::chrono::duration_cast<milliseconds>(end - start);
-				double captureTime = ms.count() / 1000;
+				double captureTime = ms.count() / 1000.0;
 				if (!frame.empty()) {
-					double scaleFactor = 1.0 / 2.0;
+					double scaleFactor;
+					if (choice == 2) {
+						scaleFactor = 1.0;
+					}
+					else {
+						scaleFactor = 1.0 / 2.0;
+					}
+
 					if (initialising) {
 						cv::putText(frame, "Initialising", cv::Point(15, 70),
 							cv::FONT_HERSHEY_SIMPLEX, 2, cv::Scalar(255, 255, 255), 2);
@@ -101,16 +122,34 @@ int main(int argc, const char** argv) {
 					else {
 						cv::putText(frame, "Recording! Try to keep still...", cv::Point(15, 70),
 							cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
-
-						char buffer[50];
-						sprintf_s(buffer, "rPPG Estimate of Heart-Rate is: %.2f", rPPG_HREst);
-						cv::putText(frame, buffer, cv::Point(15, 140),
-							cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
-						
-						char buffer2[50];
-						sprintf_s(buffer2, "rBCG Estimate of Heart-Rate is: %.2f", rBCG_HREst);
-						cv::putText(frame, buffer2, cv::Point(15, 210),
-							cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
+						if (firstRecording) {
+						}
+						else {
+							if (rPPG_HREst >= 0) {
+								char buffer[50];
+								sprintf_s(buffer, "rPPG Estimate of Heart-Rate is: %.2f", rPPG_HREst);
+								cv::putText(frame, buffer, cv::Point(15, 140),
+									cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
+							}
+							else {
+								char buffer[50];
+								sprintf_s(buffer, "Please try again - bad rPPG estimation");
+								cv::putText(frame, buffer, cv::Point(15, 140),
+									cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
+							}
+							if (rBCG_HREst >= 0) {
+								char buffer2[50];
+								sprintf_s(buffer2, "rBCG Estimate of Heart-Rate is: %.2f", rBCG_HREst);
+								cv::putText(frame, buffer2, cv::Point(15, 210),
+									cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
+							}
+							else {
+								char buffer[50];
+								sprintf_s(buffer, "Please try again - bad rBCG estimation");
+								cv::putText(frame, buffer, cv::Point(15, 210),
+									cv::FONT_HERSHEY_SIMPLEX, 1.5, cv::Scalar(255, 255, 255), 2);
+							}
+						}
 
 						Mat resizedFrame;
 						cv::resize(frame, resizedFrame, cv::Size(), scaleFactor, scaleFactor);
@@ -133,16 +172,18 @@ int main(int argc, const char** argv) {
 					timeVec.push_back(ms.count());
 				}
 
-				if (timeVec.size() > numFrames) {
+				if (timeVec.size() >= numFrames) {
 					//timeVec will continuously be replaced with newer values 
 					assert(!timeVec.empty());
 					timeVec.erase(timeVec.begin());
 				}
 			}
 			cv::destroyWindow("Live-Camera Footage");
+			
 			if (run == false) {
 				processing = false;
 				recording = false;
+				firstRecording = false;
 			}
 			else {
 				recording = false;
@@ -159,18 +200,25 @@ int main(int argc, const char** argv) {
 				initialising = false;
 			}
 			else {
+				firstRecording = false;
+
 				Rect rBCG_ROI;
 				Mat skinFrame;
 				std::vector<Rect> rBCG_ROI_Queue;
 				std::deque<Mat> skinFrameQ;
-				for (size_t j = 0; j < numFrames; j++) {
-					std::tie(skinFrame, rBCG_ROI) = detectAndDisplay(frameQ[j].clone());
-					
+				Clock::time_point start = Clock::now();
+				
+				
+#pragma omp for ordered schedule(static)
+				for (int j = 0; j < numFrames; j++) {
+
+					std::tie(skinFrame, rBCG_ROI) = detectAndDisplay(frameQ[j].clone(), choice);
+
 					if (!rBCG_ROI.empty()) {
 						rBCG_ROI_Queue.push_back(rBCG_ROI);
 					}
 					else {
-						std::cout << "rBCG_ROI is empty " << std::endl;
+						std::cout << "rBCG_ROI is empty, replacing with previous sample" << std::endl;
 						//if(j!=0) implementation in the future
 						rBCG_ROI = rBCG_ROI_Queue[j - 1];
 						rBCG_ROI_Queue.push_back(rBCG_ROI);
@@ -179,19 +227,26 @@ int main(int argc, const char** argv) {
 					if (!skinFrame.empty()) {
 						skinFrameQ.push_back(skinFrame);
 					}
-					else { 
+					else {
 						//to avoid crashing if a face is not detected
-						std::cout << "skinframe is empty " << std::endl;
+						std::cout << "skinframe is empty, replacing with previous sample" << std::endl;
 						skinFrame = skinFrameQ[j - 1];
 						skinFrameQ.push_back(skinFrame);
+
 					}
 				}
-				
+
+				Clock::time_point end = Clock::now();
+				auto ms = std::chrono::duration_cast<milliseconds>(end - start);
+				double captureTime = ms.count() / 1000.0;
+
+				std::cout << "Skin/ROI Detection Took: " << captureTime << std::endl;
+
 				// Spatial Rotation Method
 				std::vector<double> filSig_rPPG, rawSig_rPPG, rPPG_time;
 				SRResult = spatialRotation(skinFrameQ, SRResult, numFrames);
 				std::tie(rPPG_time, rawSig_rPPG, filSig_rPPG, rPPG_HREst) = rPPGEstimate(SRResult, numFrames);
-				
+
 				if (DEBUG_MODE) {
 					for (int i = 0; i < rPPG_time.size(); i++) {
 						std::cout << "time: " << rPPG_time[i] <<
@@ -200,9 +255,15 @@ int main(int argc, const char** argv) {
 					}
 				}
 
+				Clock::time_point start1 = Clock::now();
 				// KLT Optical Flow Method
 				std::vector<double> filSig_rBCG, rawSig_rBCG, rBCG_time;
 				std::tie(filSig_rBCG, rawSig_rBCG, rBCG_time, rBCG_HREst) = KLTEstimate(frameQ, rBCG_ROI_Queue, numFrames);
+
+				Clock::time_point end1 = Clock::now();
+				auto ms1 = std::chrono::duration_cast<milliseconds>(end1 - start1);
+				double captureTime2 = ms1.count() / 1000.0;
+				std::cout << "KLT Estimation took around: " << captureTime2 << " seconds " << std::endl;
 
 				if (DEBUG_MODE) {
 					for (int i = 0; i < rBCG_time.size(); i++) {
@@ -212,7 +273,7 @@ int main(int argc, const char** argv) {
 					}
 
 				}
-				
+
 				if (!filSig_rPPG.empty()) {
 					plt::subplot(2, 1, 1);
 					plt::plot(rPPG_time, rawSig_rPPG, { {"color","blue"},{"label","Raw rPPG Signal"} });
@@ -222,7 +283,10 @@ int main(int argc, const char** argv) {
 					plt::ylabel("Measured Rotation");
 
 				}
-				
+				if (plotLegends) {
+					plt::legend();
+				}
+
 				if (!filSig_rBCG.empty()) {
 					plt::subplot(2, 1, 2);
 					plt::plot(rBCG_time, rawSig_rBCG, { {"color", "red"}, {"label","Raw rBCG Signal"} });
@@ -231,13 +295,19 @@ int main(int argc, const char** argv) {
 					plt::ylabel("nominal displacement (pixels)");
 
 				}
+				if (plotLegends) {
+					plt::legend();
+					plotLegends = false;
+				}
 
 				skinFrameQ.clear(); // clear out the skinFrameQ to save RAM
 				firstStride = true; // reset the stride starting point for rPPG 
 			}
 
-			cv::destroyWindow("Detected Frame");
-
+			if (DEBUG_MODE2) {
+				cv::destroyWindow("Detected Frame");
+				cv::destroyWindow("skinFrame");
+			}
 			plt::draw();
 			plt::pause(0.001);
 			plt::save("plot.pdf");
@@ -257,7 +327,6 @@ int main(int argc, const char** argv) {
 
 	plt::draw();
 	plt::pause(0.001);
-	plt::legend();
 	plt::save("plot.pdf");
 
 	std::cout << "the program has stopped" << std::endl;
